GS-14503 implement getPropertyValue for SchemaTable - cast value to match column type

+++ xap-extensions/xap-jdbc/src/main/java/com/gigaspaces/jdbc/calcite/pg/PgCalciteTable.java
@@ -101,7 +101,7 @@ public class PgCalciteTable extends AbstractTable {
             String fqn = "public." + name;
             int oid = PgOidGenerator.INSTANCE.oid(fqn);
             ITypeDesc typeDesc = SQLUtil.checkTableExistence(name, space);
-            long idx = 0;
+            short idx = 0;
             for (PropertyInfo property : typeDesc.getProperties()) {
                 SqlTypeName sqlTypeName = GSTable.mapToSqlType(property.getType());
                 PgTypeDescriptor pgType = PgTypeUtils.fromSqlTypeName(sqlTypeName);
